% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoDude.R
\name{GMEAN.MM}
\alias{GMEAN.MM}
\title{GMEAN.MM}
\usage{
GMEAN.MM(m, V)
}
\arguments{
\item{m}{The mean of the data, an estimator for \eqn{E[x]}.}

\item{V}{The variance of the data, an estimator for \eqn{E[x^2] - E[x]^2}.}
}
\value{
GMEAN.MM
}
\description{
This function calculates Geometric Mean using the Method of Moments, as well as
the Geometric SD. The only inputs are the mean and variance of the data. The data is assumed to be log-normal.
\deqn{\textrm{GMEAN}(m,V) = \exp{(\mu_l)}; ~\mu_l = 2 \ln(m) - \ln \left (\sqrt{V + m^2} \right )}
\deqn{\textrm{GSD}(m,V) = \exp(\sigma_l); ~\sigma_l = \sqrt{\ln(1.0 + V/m^2)}}
}
\examples{
GMEAN.MM( mean( 9:20 * 10 ), var((9:20)*10 ) ) ## Returns $g.mean 140.715, $g.sd 1.277543
}
\keyword{GMEAN.MM}
