% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoDude.R
\name{GSEM}
\alias{GSEM}
\title{GSEM}
\usage{
GSEM(x)
}
\arguments{
\item{x}{Data points, as a vector-like object}
}
\value{
GSEM
}
\description{
This function calculates Geometric SEM from a set of data
\deqn{\textrm{GSEM}(x_i) = \exp{(sem_l)};~ sem_l = \frac{\sigma_l}{\sqrt{n}}}
}
\examples{
GSEM((9:20)*10) ## Returns 1.077717
}
\keyword{GSEM}
