% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoDude.R
\name{MOMENTS.MM}
\alias{MOMENTS.MM}
\title{MOMENTS.MM}
\usage{
MOMENTS.MM(gmean, gsd)
}
\arguments{
\item{gmean}{The geometric mean of the data.}

\item{gsd}{The geometric SD.}
}
\value{
MOMENTS.MM
}
\description{
This function calculates arithmetic Mean as well as the variance from the Geometric mean and sd using the Method of Moments. The only inputs are the geometric mean and variance of the data. The data is assumed to be log-normal.
\deqn{\textrm{GMEAN}(m,V) = \exp{(\mu_l)}; ~\mu_l = 2 \ln(m) - \ln \left (\sqrt{V + m^2} \right )}
\deqn{\textrm{GSD}(m,V) = \exp(\sigma_l); ~\sigma_l = \sqrt{\ln(1.0 + V/m^2)}}
}
\examples{
MOMENTS.MM( 140.0, 1.3 ) ## Returns $mean 144.9023, $var 1496.212
}
\keyword{MOMENTS.MM}
